Updates due to comments, refine styles
