+++ src/screens/AddEditObservation/sections/PhotoCarousel.js
@@ -1,84 +1,15 @@
 import React, { useState } from "react";
-import {
-  TouchableHighlight,
-  TouchableOpacity,
-  Image,
-  ScrollView,
-  Text,
-  View
-} from "react-native";
+import { TouchableHighlight, Image, ScrollView } from "react-native";
 import PropTypes from "prop-types";
-import ImagePicker from "react-native-image-picker";
-import { Overlay } from "react-native-elements";
 
-import { Button } from "../../../components";
-import { translate } from "../../../i18n";
+import PhotoResized from "./PhotoResized";
+import PhotoChooseWindow from "./PhotoChooseWindow";
 import { styles } from "../styles";
 
 const photoPlaceholder = require("../../../assets/photoPlaceholder.png");
 
-const PhotoChooseWindow = props => {
-  const { onPhotoChosen, onCloseHandler } = props;
-
-  const onOpenGallery = () => {
-    ImagePicker.launchImageLibrary({}, onPhotoChosen);
-  };
-
-  const onTakePhoto = () => {
-    ImagePicker.launchCamera({}, onPhotoChosen);
-  };
-
-  const onClose = () => {
-    onCloseHandler();
-  };
-
-  return (
-    <Overlay
-      height={400}
-      width={280}
-      borderRadius={6}
-      contentContainerStyle={styles.overlay}
-    >
-      <Text style={styles.overlayTitle}>
-        {translate("editObservation.addPhotoTitle")}
-      </Text>
-      <Text style={styles.overlayDescription}>
-        {translate("editObservation.addPhotoDescription")}
-      </Text>
-      <Button
-        wrapperStyles={styles.overlayGallery}
-        onPress={onOpenGallery}
-        caption={translate("editObservation.gallery")}
-        appearance="Dark"
-      />
-      <Button
-        wrapperStyles={styles.overlayTakePhoto}
-        onPress={onTakePhoto}
-        caption={translate("editObservation.makePhoto")}
-        appearance="Light"
-      />
-      <View style={styles.overlayCloseWindowContainer}>
-        <Button
-          wrapperStyles={styles.overlayCloseWindow}
-          onPress={onClose}
-          caption={translate("editObservation.close")}
-          appearance="Borderless"
-        />
-      </View>
-    </Overlay>
-  );
-};
-PhotoChooseWindow.propTypes = {
-  onPhotoChosen: PropTypes.func,
-  onCloseHandler: PropTypes.func
-};
-PhotoChooseWindow.defaultProps = {
-  onPhotoChosen: () => {},
-  onCloseHandler: () => {}
-};
-
 const PhotoCarousel = props => {
-  const { photos, updateFieldValue, birdPhotos, setFieldValue } = props;
+  const { updateFieldValue, birdPhotos, setFieldValue } = props;
 
   const [isUploadPhotoVisible, setUploadPhotoVisible] = useState(false);
   const [isPhotoResized, setPhotoResized] = useState(false);
@@ -109,7 +40,9 @@ const PhotoCarousel = props => {
   };
   const onDeletePhoto = () => {
     setPhotoResized(false);
-    const filteredPhotos = photos.filter(photo => photo !== photoForResizing);
+    const filteredPhotos = birdPhotos.filter(
+      photo => photo !== photoForResizing
+    );
     setFieldValue({ birdPhotos: filteredPhotos });
   };
 
@@ -119,36 +52,20 @@ const PhotoCarousel = props => {
       horizontal
       contentContainerStyle={styles.photosContainer}
     >
-      <Overlay
-        isVisible={isPhotoResized}
-        overlayBackgroundColor="rgba(0, 0, 0, 0.6)"
-        fullScreen
-      >
-        <TouchableOpacity
-          onPress={onCloseResizedPhoto}
-          style={styles.resizedPhotoButton}
-        >
-          <Text style={styles.resizedPhotoButtonText}>
-            {translate("editObservation.close")}
-          </Text>
-        </TouchableOpacity>
-        <Image style={styles.birdPhotoResized} source={photoForResizing} />
-        <TouchableOpacity
-          onPress={onDeletePhoto}
-          style={styles.resizedPhotoButton}
-        >
-          <Text style={styles.resizedPhotoButtonText}>
-            {translate("editObservation.delete")}
-          </Text>
-        </TouchableOpacity>
-      </Overlay>
+      {isPhotoResized ? (
+        <PhotoResized
+          onCloseResizedPhoto={onCloseResizedPhoto}
+          onDeletePhoto={onDeletePhoto}
+          photoForResizing={photoForResizing}
+        />
+      ) : null}
       {isUploadPhotoVisible ? (
         <PhotoChooseWindow
           onPhotoChosen={onPhotoChosen}
           onCloseHandler={onChoosePhotoWindowClose}
         />
       ) : null}
-      {photos.map(photoPath => (
+      {birdPhotos.map(photoPath => (
         <TouchableHighlight
           key={photoPath}
           onPress={() => resizePhoto(photoPath)}
@@ -164,13 +81,11 @@ const PhotoCarousel = props => {
 };
 
 PhotoCarousel.propTypes = {
-  photos: PropTypes.array,
   updateFieldValue: PropTypes.func.isRequired,
   birdPhotos: PropTypes.array,
   setFieldValue: PropTypes.func
 };
 PhotoCarousel.defaultProps = {
-  photos: [],
   birdPhotos: [],
   setFieldValue: () => {}
 };
